minifridge_gcloud/functions/requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
idna==2.10
imutils==0.5.3
numpy==1.19.1
opencv-python==4.4.0.42
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
